Merge the aeros-qt4 branch into the trunk, details below.
Trunk changes from 4101 to 4111 had already been merged into the branch, so I first
merged 4111-4329 from the trunk/source into the branch.

 svn merge -r4111:4329 http://svn/svn/aeros/trunk/source .

So the branch then contained all the changes made on the trunk since the
branch was created, minus a few conflicts in MainWindow.cc which had become
obsolete.

Next checked out the trunk source, and reverted all changes in the trunk
since the branch:

 svn merge -r4330:4090 .

Finally merged in all the changes on the branch.  The second merge command
picked up changes Charlie made to the branch after the first one.

 svn merge -r4090:4330 http://svn/svn/aeros/branches/webster/aeros-qt4 
 svn merge -r4330:4331 http://svn/svn/aeros/branches/webster/aeros-qt4

As expected, these merges succeed without any conflicts.  A diff between
the latest branch source and trunk source shows no differences:

 diff --recursive --exclude=.svn source-qt4-snoopy source

This is the copy being committed.  See the branch for log details.

This wiki page has some info on installing the new build requirements for
this version of aeros:

 http://wiki.eol.ucar.edu/sew/AerosQt4Sprint