fix CID#10086: use float operand for float math
I think the intention here was to allow a floating point result for the
radius, since this is SVG after all, and otherwise symbol sizes of 4 and 5
will look the same.  However, fixing this changes all the ASPEN SVG output.
Therefore, this also changes the default symbol size to 4, since that gives
the same radius as before.
Also fix some unitialized scalar class members.